+++ packages/@expo/cli/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- Lookup absolute asset paths correctly when used as server components. ([#36501](https://github.com/expo/expo/pull/36501) by [@byCedric](https://github.com/byCedric))
+
 ### 💡 Others
 
 ## 0.24.6 — 2025-04-28